requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.1
langchain==0.3.26
transformers
huggingface-hub
accelerate
scipy==1.16.0
scikit-learn==1.7.1
google-generativeai
python-multipart
numpy==2.3.1
pydantic==2.11.7
starlette==0.47.2
langchain
langchain-google-genai==2.0.10
streamlit==1.47.0
requests==2.32.4
python-dotenv==1.1.1
protobuf==5.29.5
langchain-experimental==0.3.4